backend/requirements.txt
fastapi
google-genai==1.13.0
pydantic
pymongo
pillow
pydantic_settings
python-multipart
uvicorn
